Refactoring some KDF related functions
+ The refactored methods are common between HKDF, concatenation KDF, and
  CTR KDF.
